fix: tasks

+++ src/operations/file.js
@@ -1,31 +1,24 @@
-import { isAbsolute, resolve } from "path";
 import fs from "fs";
 import process from "process";
-
-let file;
-
-const checkFileExist = (path, currentDir) => {
-  if (isAbsolute(path)) {
-    file = path;
-  } else {
-    const currentPath = resolve(currentDir, path);
-    file = currentPath;
-  }
-};
+import { checkFileExist, file, isValidPath } from "./general.js";
 
 export async function cat(path, currentDir) {
   checkFileExist(path, currentDir);
-  let reader = new fs.createReadStream(file);
+
+  let reader = fs.createReadStream(file);
   reader.pipe(process.stdout);
   const checkErrors = new Promise((resolve, reject) => {
     reader.on("end", () => resolve());
-    reader.on("error", () => console.log("File not read"));
+    reader.on("error", () =>
+      console.log("Error: No such file exists and File not read")
+    );
   });
   await checkErrors;
 }
 
 export async function add(path, currentDir) {
   checkFileExist(path, currentDir);
+
   let writer = fs.createWriteStream(file);
   const checkErrors = new Promise((resolve, reject) => {
     writer.on("end", () => resolve());
@@ -37,43 +30,60 @@ export async function add(path, currentDir) {
 export function rn(path, currentDir, newFile) {
   checkFileExist(path, currentDir);
 
-  if (fs.existsSync(file) && !fs.existsSync(newFile)) {
-    fs.rename(file, newFile, () => {
-      console.log("File Renamed!");
-    });
+  if (fs.existsSync(isValidPath(currentDir, newFile))) {
+    if (fs.existsSync(file) && !fs.existsSync(newFile)) {
+      fs.rename(file, newFile, () => {
+        console.log("File Renamed!");
+      });
+    } else {
+      console.log("Error: No such file exists and File not rename");
+    }
   } else {
-    const err = new Error("FS operation failed");
-    return console.log("File not rename");
+    console.log("Error: New file directory not found");
   }
 }
 
 export async function cp(path, currentDir, newFile, unlink) {
   checkFileExist(path, currentDir);
-  if (newFile.includes(".txt")) {
-    let reader = new fs.createReadStream(file, "utf-8");
-    let writer = fs.createWriteStream(newFile);
 
-    reader.on("data", function (chunk) {
-      writer.write(chunk);
-      console.log("Completed successfully");
-    });
+  if (fs.existsSync(isValidPath(currentDir, newFile))) {
+    if (
+      fs.existsSync(file) &&
+      file.includes(".txt") &&
+      newFile.includes(".txt")
+    ) {
+      let reader = new fs.createReadStream(file, "utf-8");
+      let writer = fs.createWriteStream(newFile);
 
-    const checkErrorsReader = new Promise((resolve, reject) => {
-      reader.on("end", () => resolve());
-      reader.on("error", () => console.log("File not read"));
-      writer.on("end", () => resolve());
-      writer.on("error", () => console.log("File not created"));
-    });
-    await checkErrorsReader;
-    if (unlink) {
-      try {
-        fs.unlinkSync(file);
-      } catch (error) {
-        console.log("err");
+      reader.on("data", function (chunk) {
+        writer.write(chunk);
+      });
+
+      const checkErrorsReader = new Promise((resolve, reject) => {
+        reader.on("end", () => resolve());
+        reader.on("error", () =>
+          console.log("Error: No such file exists and File not read")
+        );
+        writer.on("end", () => resolve());
+        writer.on("error", () =>
+          console.log("Error: No such file exists and File not created")
+        );
+      });
+      await checkErrorsReader;
+      if (unlink) {
+        try {
+          fs.unlinkSync(file);
+        } catch (error) {
+          console.log("Error: File not read an not created");
+        }
       }
+    } else {
+      console.log(
+        `Error: No such file exists and The file and new file must include ".txt"`
+      );
     }
   } else {
-    console.log(`The new file must include ".txt"`);
+    console.log("Error: New file directory not found");
   }
 }
 
@@ -89,6 +99,6 @@ export async function rm(path, currentDir) {
     fs.unlinkSync(file);
     console.log("File deleted");
   } catch (error) {
-    console.log("No such file exists");
+    console.log("Error: No such file exists");
   }
 }
